+++ Backend/.env
@@ -1,5 +1,5 @@
 DB_HOST=my-donation-db.cjm6ce8cikky.us-east-1.rds.amazonaws.com
 DB_USER=admin
 DB_PASSWORD=qBDiSihbNAICSeX0RR8P
-DB_NAME=donations
+DB_NAME=Newsletter
 PORT=3001
